feat(codeql): Update Pack support
