+++ core/src/codeql/packs/mod.rs
@@ -2,8 +2,10 @@
 
 pub mod handler;
 pub mod loader;
+pub mod models;
 pub mod pack;
 pub mod packs;
 
-pub use pack::{CodeQLPack, CodeQLPackType};
+pub use models::{CodeQLPackType, PackYaml, PackYamlLock};
+pub use pack::CodeQLPack;
 pub use packs::CodeQLPacks;
